requirements.txt
# Core web framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# Templating
jinja2>=3.1.4

# Form parsing / multipart uploads
python-multipart>=0.0.9

# Environment variables
python-dotenv>=1.0.0

# Data validation
pydantic>=2.7.0
pydantic-settings>=2.3.0

# LLM
openai>=1.30.0

# Vector database
chromadb>=0.5.0

# Document reading
pymupdf>=1.24.0        # PDF support (fast)
pypdf>=4.2.0           # PDF fallback
python-docx>=1.1.0     # DOCX support

# Authentication
authlib>=1.3.0         # Google OAuth 2.0 flow
itsdangerous>=2.1.0    # Starlette SessionMiddleware cookie signing

# Utilities
httpx>=0.27.0          # async HTTP (used internally by FastAPI)
huggingface_hub>=0.23.0  # HF Dataset sync for persistent storage on HF Spaces

# Billing / QR code
qrcode[pil]>=7.4.2     # UPI QR code generation (pulls in Pillow)
